Move step1 to its own file
