+++ mlir/include/mlir/Dialect/MIOpen/LowerMIOpenOps.h
@@ -22,6 +22,7 @@
 #include "mlir/Dialect/LLVMIR/LLVMDialect.h"
 #include "mlir/Dialect/MIOpen/AffineMapHelper.h"
 #include "mlir/Dialect/MIOpen/Passes.h"
+#include "mlir/Dialect/MIOpen/Tuning/GridwiseGemmParams.h"
 #include "mlir/Dialect/MemRef/IR/MemRef.h"
 #include "mlir/Dialect/SCF/SCF.h"
 #include "mlir/Dialect/StandardOps/IR/Ops.h"
@@ -46,7 +47,7 @@
 #include "mlir/Transforms/Passes.h"
 
 #include "XdlopsCodeSelection.h"
-#include "mlir/Dialect/MIOpen/Tuning/GridwiseGemmParams.h"
+
 #include "mlir/Dialect/MIOpen/utility/common.h"
 #include "mlir/Dialect/MIOpen/utility/loweringUtils.h"
 #include "mlir/Dialect/MIOpen/utility/math.h"
@@ -1339,677 +1340,6 @@ inline void computeBottomIndicesWithAffineMap(
   }
 }
 
-//===----------------------------------------------------------------------===//
-// Conv2D (forward, backward) lowering.
-//===----------------------------------------------------------------------===//
-
-// Forward declare specialized lowerings that've been moved elsewhere
-LogicalResult backwardWeightAtomicAdd(miopen::Conv2DBwdWeightOp op,
-                                      PatternRewriter &b);
-LogicalResult backwardData(miopen::Conv2DBwdDataOp op, PatternRewriter &b);
-template <typename T>
-LogicalResult checkNames(ArrayRef<StringRef> actual,
-                         ArrayRef<StringRef> expected, StringRef argName,
-                         T op) {
-  if (actual.size() != expected.size()) {
-    return op.emitOpError("Layout mismatch in ")
-           << argName << " tensor: Expected " << expected.size()
-           << " dimensions but have " << actual.size();
-  }
-  for (StringRef name : expected) {
-    if (std::find(actual.begin(), actual.end(), name) == actual.end()) {
-      return op.emitOpError("Layout mismatch in ")
-             << argName << " tensor: Expected it to have a `" << name
-             << "` dimension";
-    }
-  }
-  return success();
-}
-
-template <typename T> struct Conv2DRewritePattern : public OpRewritePattern<T> {
-  const static ArgumentFields fields;
-  const static miopen::ConvOpType convOpType;
-  using OpRewritePattern<T>::OpRewritePattern;
-
-  LogicalResult matchAndRewrite(T op, PatternRewriter &b) const override {
-    bool isXdlops = false;
-    auto xdlopsV2Attr = op->template getAttrOfType<BoolAttr>("xdlopsV2");
-    if (xdlopsV2Attr && xdlopsV2Attr.getValue() == true)
-      isXdlops = true;
-    auto dataType =
-        op.input().getType().template cast<MemRefType>().getElementType();
-    if (miopen::ConvOpType::BwdData == convOpType) {
-      return backwardData(cast<miopen::Conv2DBwdDataOp>(op), b);
-    }
-    auto loc = op.getLoc();
-
-    auto archAttr = op->template getAttrOfType<StringAttr>("arch");
-    auto numCuAttr = op->template getAttrOfType<IntegerAttr>("num_cu");
-
-    auto filterLayoutAttr =
-        op->template getAttrOfType<ArrayAttr>("filter_layout");
-    auto inputLayoutAttr =
-        op->template getAttrOfType<ArrayAttr>("input_layout");
-    auto outputLayoutAttr =
-        op->template getAttrOfType<ArrayAttr>("output_layout");
-
-    auto dilationsAttr = op->template getAttrOfType<ArrayAttr>("dilations");
-    auto stridesAttr = op->template getAttrOfType<ArrayAttr>("strides");
-    auto paddingAttr = op->template getAttrOfType<ArrayAttr>("padding");
-
-    // Get shape of filter tensor.
-    auto filterType = op.filter().getType().template cast<MemRefType>();
-    auto filterShape = filterType.getShape();
-
-    // Get shape of input tensor.
-    auto inputType = op.input().getType().template cast<MemRefType>();
-    auto inputShape = inputType.getShape();
-
-    // Get shape of output tensor.
-    auto outputType = op.output().getType().template cast<MemRefType>();
-    auto outputShape = outputType.getShape();
-
-    // Obtain convolution parameters: padding / dialtion / stride.
-    int64_t leftPadH =
-        paddingAttr.getValue()[0].template cast<IntegerAttr>().getInt();
-    int64_t leftPadW =
-        paddingAttr.getValue()[2].template cast<IntegerAttr>().getInt();
-    int64_t rightPadH =
-        paddingAttr.getValue()[1].template cast<IntegerAttr>().getInt();
-    int64_t rightPadW =
-        paddingAttr.getValue()[3].template cast<IntegerAttr>().getInt();
-
-    int64_t dilationH =
-        dilationsAttr.getValue()[0].template cast<IntegerAttr>().getInt();
-    int64_t dilationW =
-        dilationsAttr.getValue()[1].template cast<IntegerAttr>().getInt();
-    int64_t strideH =
-        stridesAttr.getValue()[0].template cast<IntegerAttr>().getInt();
-    int64_t strideW =
-        stridesAttr.getValue()[1].template cast<IntegerAttr>().getInt();
-
-    // get y, x, ho, wo, hi, wi, k, c, n
-    int64_t y, x, ho, wo, hi, wi, k, c, n;
-    y = x = ho = wo = hi = wi = k = c = n = 0;
-    llvm::SmallVector<StringRef, 5> filterNames, inputNames, outputNames;
-    for (uint32_t i = 0; i < filterLayoutAttr.size(); ++i) {
-      auto filterAttr =
-          filterLayoutAttr.getValue()[i].template cast<StringAttr>();
-      auto inputAttr =
-          inputLayoutAttr.getValue()[i].template cast<StringAttr>();
-      auto outputAttr =
-          outputLayoutAttr.getValue()[i].template cast<StringAttr>();
-
-      filterNames.push_back(filterAttr.getValue());
-      inputNames.push_back(inputAttr.getValue());
-      outputNames.push_back(outputAttr.getValue());
-
-      if (filterAttr.getValue() == "y") {
-        y = filterShape[i];
-      } else if (filterAttr.getValue() == "x") {
-        x = filterShape[i];
-      } else if (filterAttr.getValue() == "k") {
-        k = filterShape[i];
-      } else if (filterAttr.getValue() == "c") {
-        c = filterShape[i];
-      }
-
-      if (inputAttr.getValue() == "hi") {
-        hi = inputShape[i];
-      } else if (inputAttr.getValue() == "wi") {
-        wi = inputShape[i];
-      } else if (inputAttr.getValue() == "ni") {
-        n = inputShape[i];
-      }
-
-      if (outputAttr.getValue() == "ho") {
-        ho = outputShape[i];
-      } else if (outputAttr.getValue() == "wo") {
-        wo = outputShape[i];
-      }
-    }
-
-    if (failed(
-            checkNames(filterNames, {"k", "g", "c", "y", "x"}, "filter", op)) ||
-        failed(checkNames(inputNames, {"ni", "gi", "ci", "hi", "wi"}, "input",
-                          op)) ||
-        failed(checkNames(outputNames, {"no", "go", "ko", "ho", "wo"}, "output",
-                          op))) {
-      return failure();
-    }
-
-    int64_t gemmMSize, gemmNSize, gemmKSize;
-    int64_t gemmMExtra, gemmNExtra, gemmKExtra;
-    gemmMSize = gemmNSize = gemmKSize = 0;
-    gemmMExtra = gemmNExtra = gemmKExtra = 0;
-    // compute we should use extra padding kernel or not
-    // c,k already / g ,so we can skip / g here
-    switch (convOpType) {
-    case miopen::ConvOpType::Fwd:
-      gemmMSize = k;
-      gemmKSize = c * y * x;
-      gemmNSize = n * ho * wo;
-      break;
-    case miopen::ConvOpType::BwdData:
-      gemmMSize = c;
-      gemmKSize = k * y * x;
-      gemmNSize = n * ho * wo;
-      break;
-    case miopen::ConvOpType::BwdWeight:
-      gemmMSize = k;
-      gemmKSize = n * ho * wo;
-      gemmNSize = c * y * x;
-      break;
-    }
-
-    bool needExtraPad = false;
-
-    auto calculatePaddingKernelSize = [&needExtraPad, gemmMSize, gemmNSize,
-                                       gemmKSize, &gemmMExtra, &gemmNExtra,
-                                       &gemmKExtra](auto populateParams) {
-      auto configParams = populateParams.getTuningParameters();
-      size_t numOfFailedConfigs = 0;
-      for (auto &params : configParams) {
-        if (gemmMSize % params.gemmMPerBlock == 0 &&
-            gemmKSize % params.gemmKPerBlock == 0 &&
-            gemmNSize % params.gemmNPerBlock == 0) {
-          break;
-        }
-        numOfFailedConfigs++;
-      }
-
-      auto extraParams = populateParams.getUniversalParameters();
-      if (numOfFailedConfigs == configParams.size()) {
-        needExtraPad = true;
-        int64_t gemmMRemain, gemmKRemain, gemmNRemain;
-
-        gemmMRemain = gemmMSize % extraParams.gemmMPerBlock;
-        if (gemmMRemain != 0)
-          gemmMExtra = extraParams.gemmMPerBlock - gemmMRemain;
-
-        gemmNRemain = gemmNSize % extraParams.gemmNPerBlock;
-        if (gemmNRemain != 0)
-          gemmNExtra = extraParams.gemmNPerBlock - gemmNRemain;
-
-        gemmKRemain = gemmKSize % extraParams.gemmKPerBlock;
-        if (gemmKRemain != 0)
-          gemmKExtra = extraParams.gemmKPerBlock - gemmKRemain;
-
-        // llvm::errs() << "gemmMExtra: " << gemmMExtra << "gemmNExtra: " <<
-        // gemmNExtra << "gemmKExtra: " << gemmKExtra << "\n";
-      }
-    };
-
-    if (!isXdlops) {
-      PopulateParams populateParams;
-      calculatePaddingKernelSize(populateParams);
-    } else { // xdlops
-      PopulateParamsXDL populateParamsXDL;
-      calculatePaddingKernelSize(populateParamsXDL);
-    }
-
-    if (miopen::ConvOpType::BwdWeight == convOpType && isXdlops &&
-        dataType == b.getF32Type() && needExtraPad == false) {
-      // current backward weight with atomic_add can only run under xdlops +
-      // fp32
-      return backwardWeightAtomicAdd(cast<miopen::Conv2DBwdWeightOp>(op), b);
-    }
-
-    // Transform filter tensor.
-
-    // filter dims need oob check
-    OobCheckSet filterOobCheckDims;
-    // set layout attribute.
-    // Weight tensor transformation for Conv2DOp
-    // - PassThrough G dimension to dimension 0, name it gemmG.
-    // - Merge non-K dimensions to dimension 1, name it as gemmK.
-    //   Optimization: If non-K dimensions are consequetive, apply unfold.
-    // - PassThrough K dimension to dimension 2, name it as gemmM.
-    //
-    // Weight tensor transformation for Conv2DBwdWeightOp
-    // - PassThrough G dimension to dimension 0, name it gemmG
-    // - PassThrough K dimension to dimension 1, name it as gemmM.
-    // - Merge non-K dimensions to dimension 2, name it as gemmN.
-    SmallVector<StringRef, 5> filterNonKDims;
-    for (StringRef name : filterNames)
-      if (name != "g" && name != "k")
-        filterNonKDims.push_back(name);
-
-    BottomUpCTBuilder filterTransform(b, filterNames, filterShape, loc);
-    filterTransform.passThrough({"gemmG"}, {0}, {"g"});
-    bool isUnfold = filterTransform.startIndex("g") == 0 &&
-                    (filterTransform.startIndex("k") == 1 ||
-                     filterTransform.startIndex("k") == 4);
-    switch (convOpType) {
-    case miopen::ConvOpType::Fwd:
-      filterTransform.merge("gemmK", 1, filterNonKDims, isUnfold);
-      filterTransform.passThrough({"gemmM"}, {2}, {"k"});
-      break;
-    case miopen::ConvOpType::BwdWeight:
-      filterTransform.passThrough({"gemmM"}, {1}, {"k"});
-      filterTransform.merge("gemmN", 2, filterNonKDims, isUnfold);
-      break;
-    case miopen::ConvOpType::BwdData:
-      llvm_unreachable("Backward data has been sent elsewhere");
-      break;
-    }
-
-    TransformMapAttr filterTransformAttr = filterTransform.get();
-    Value gemmFilter =
-        b.create<miopen::TransformOp>(loc, op.filter(), filterTransformAttr);
-
-    Value gemmFilterPad = gemmFilter;
-
-    // filter pad start
-    // K:output channel, C:input channel,Y:filter height,X:filter width
-    // filter dim : K & merge(C,Y,X) , if C*Y*X is under 64 or 32
-    // we pad CYX to 32 or 64, then mlir can do gemm
-    // we add more one transform to do pad
-    bool filterCheckPadGemmM = false;
-    bool filterCheckPadGemmK = false;
-    bool filterCheckPadGemmN = false;
-    filterCheckPadGemmM =
-        (convOpType == miopen::ConvOpType::Fwd && gemmMExtra > 0) ||
-        (convOpType == miopen::ConvOpType::BwdWeight && gemmMExtra > 0);
-    filterCheckPadGemmK =
-        (convOpType == miopen::ConvOpType::Fwd && gemmKExtra > 0);
-    filterCheckPadGemmN =
-        (convOpType == miopen::ConvOpType::BwdWeight && gemmNExtra > 0);
-    if (filterCheckPadGemmM || filterCheckPadGemmK || filterCheckPadGemmN) {
-      auto padGemmFilterTransform =
-          BottomUpCTBuilder::above(filterTransform, filterTransformAttr);
-      padGemmFilterTransform.passThrough("gemmG");
-
-      // Note that, when padding a gemm dimension that came from the non-K
-      // tensor dimensions, only the leading dimension is added to the oob check
-      // set, as adding all the dimensions historically led to miscompilation
-      if (filterCheckPadGemmK) {
-        padGemmFilterTransform.pad("gemmKPad", "gemmK", 0, gemmKExtra);
-        filterOobCheckDims.insert(
-            filterTransform.startIndex(filterNonKDims[0]));
-      } else if (convOpType != miopen::ConvOpType::BwdWeight) {
-        // Backward weight has no GemmK on its filter
-        padGemmFilterTransform.passThrough("gemmK");
-      }
-
-      if (filterCheckPadGemmM) {
-        padGemmFilterTransform.pad("gemmMPad", "gemmM", 0, gemmMExtra);
-        filterOobCheckDims.insert(filterTransform.startIndex("k"));
-      } else {
-        padGemmFilterTransform.passThrough("gemmM");
-      }
-
-      if (filterCheckPadGemmN) {
-        padGemmFilterTransform.pad("gemmNPad", "gemmN", 0, gemmNExtra);
-        filterOobCheckDims.insert(
-            filterTransform.startIndex(filterNonKDims[0]));
-      } else if (convOpType != miopen::ConvOpType::Fwd) {
-        padGemmFilterTransform.passThrough("gemmN");
-      }
-
-      TransformMapAttr padGemmFilterTransformAttr =
-          padGemmFilterTransform.get();
-      gemmFilterPad = b.create<miopen::TransformOp>(loc, gemmFilter,
-                                                    padGemmFilterTransformAttr);
-      // filter pad end
-    }
-
-    // Transform input tensor.
-    // Input tensor step 1: padded input.
-
-    OobCheckSet inputOobCheckDims;
-    // set layout attribute.
-    // Padded input tensor transformation:
-    // - Pass through ni, gi, and ci, not renaming them
-    // - Padd hi and wi as specified in padding attributes, renaming them to
-    // hipad and wipad
-    BottomUpCTBuilder padInputTransform(b, inputNames, inputShape, loc);
-    padInputTransform.passThrough("ni");
-    padInputTransform.passThrough("gi");
-    padInputTransform.passThrough("ci");
-
-    llvm::SmallVector<int64_t, 4> padArgs = {leftPadH, rightPadH, leftPadW,
-                                             rightPadW};
-    llvm::SmallVector<uint32_t, 2> padOutDims = {
-        padInputTransform.startIndex("hi"), padInputTransform.startIndex("wi")};
-    padInputTransform.pad({"hipad", "wipad"}, padOutDims, {"hi", "wi"},
-                          padArgs);
-    if (leftPadH || rightPadH) {
-      inputOobCheckDims.insert(padInputTransform.startIndex("hi"));
-    }
-    if (leftPadW || rightPadW) {
-      inputOobCheckDims.insert(padInputTransform.startIndex("wi"));
-    }
-
-    TransformMapAttr padInputTransformAttr = padInputTransform.get();
-
-    Value paddedInput =
-        b.create<miopen::TransformOp>(loc, op.input(), padInputTransformAttr);
-
-    // Input tensor step 2 : embedded input.
-    // Embedded input tensor transformation:
-    // - PassThrough gi, ni, and ci
-    // - Embed hipad to y and ho with size filter y by output h and
-    //   coefficients dilationH and strideH
-    // - Embed wipad to x and wo with size filter x by output h and
-    //   coefficients dilationW and strideW
-
-    llvm::SmallVector<StringRef, 5> paddedInputNames;
-    padInputTransform.getEndNames(paddedInputNames);
-    llvm::StringMap<uint32_t> embeddedInputDims;
-    for (uint32_t i = 0, newDim = 0, e = paddedInputNames.size(); i < e; ++i) {
-      StringRef name = paddedInputNames[i];
-      if (name == "hipad") {
-        embeddedInputDims.insert({"y", newDim++});
-        embeddedInputDims.insert({"ho", newDim++});
-      } else if (name == "wipad") {
-        embeddedInputDims.insert({"x", newDim++});
-        embeddedInputDims.insert({"wo", newDim++});
-      } else {
-        embeddedInputDims.insert({name, newDim++});
-      }
-    }
-
-    BottomUpCTBuilder embedInputTransform(
-        b, paddedInputNames, padInputTransformAttr.getUpperBounds(), loc);
-    embedInputTransform.passThrough({"ni", "gi", "ci"},
-                                    {embeddedInputDims["ni"],
-                                     embeddedInputDims["gi"],
-                                     embeddedInputDims["ci"]},
-                                    {"ni", "gi", "ci"});
-    embedInputTransform.embed({"y", "ho"},
-                              {embeddedInputDims["y"], embeddedInputDims["ho"]},
-                              {y, ho}, "hipad", {dilationH, strideH});
-    embedInputTransform.embed({"x", "wo"},
-                              {embeddedInputDims["x"], embeddedInputDims["wo"]},
-                              {x, wo}, "wipad", {dilationW, strideW});
-
-    TransformMapAttr embedInputTransformAttr = embedInputTransform.get();
-    Value embeddedInput = b.create<miopen::TransformOp>(
-        loc, paddedInput, embedInputTransformAttr);
-
-    // Input tensor step 3: GEMM'd input
-    //
-    // - PassThrough gi to dimension 0 and name it gemmG, then
-    // For Conv2DOp:
-    // - Merge ci, y, x dimensions to dimension 1, name it as gemmK.
-    // - Merge ni, ho, wo dimensions to dimension 2, name it as gemmN.
-    //
-    // For Conv2DBwdWeightOp:
-    // - Part 1: Merge ni, ho, wo dimensions to dimension 1, name it as gemmK.
-    // - Part 2: Merge ci, y, x dimensions to dimension 2, name it as gemmN.
-
-    auto gemmInputTransform =
-        BottomUpCTBuilder::above(embedInputTransform, embedInputTransformAttr);
-    gemmInputTransform.passThrough({"gemmG"}, {0}, {"gi"});
-
-    llvm::SmallVector<StringRef, 3> nonNHWDims = {"ci", "y", "x"};
-    std::sort(nonNHWDims.begin(), nonNHWDims.end(),
-              [&gemmInputTransform](const StringRef &v1,
-                                    const StringRef &v2) -> bool {
-                return gemmInputTransform.startIndex(v1) <
-                       gemmInputTransform.startIndex(v2);
-              });
-
-    llvm::SmallVector<StringRef, 3> mergeToK, mergeToN;
-    switch (convOpType) {
-    case miopen::ConvOpType::Fwd:
-      mergeToK = std::move(nonNHWDims);
-      mergeToN = {"ni", "ho", "wo"};
-      break;
-    case miopen::ConvOpType::BwdWeight:
-      mergeToK = {"ni", "ho", "wo"};
-      mergeToN = std::move(nonNHWDims);
-      break;
-    case miopen::ConvOpType::BwdData:
-      llvm_unreachable("Backward data is in another function");
-    }
-    gemmInputTransform.merge("gemmK", 1, mergeToK);
-    gemmInputTransform.merge("gemmN", 2, mergeToN);
-
-    TransformMapAttr gemmInputTransformAttr = gemmInputTransform.get();
-    Value gemmInput = b.create<miopen::TransformOp>(loc, embeddedInput,
-                                                    gemmInputTransformAttr);
-    Value gemmInputPad = gemmInput;
-
-    // input padding start
-    // input : NHW & CRS , if CRS is under 64 or 32
-    // we pad CRS to 32 or 64, then mlir can do gemm
-    // we add more one transform to do pad
-
-    // input forward : gemmK,gemmN
-    // backward weights: gemmK,gemmN
-    // so we don't need to pad gemmK
-    bool inputCheckPadGemmK = false;
-    bool inputCheckPadGemmN = false;
-    inputCheckPadGemmK =
-        (convOpType == miopen::ConvOpType::Fwd && gemmKExtra > 0) ||
-        (convOpType == miopen::ConvOpType::BwdWeight && gemmKExtra > 0);
-    inputCheckPadGemmN =
-        (convOpType == miopen::ConvOpType::Fwd && gemmNExtra > 0) ||
-        (convOpType == miopen::ConvOpType::BwdWeight && gemmNExtra > 0);
-    if (inputCheckPadGemmK || inputCheckPadGemmN) {
-      auto padGemmInputTransform =
-          BottomUpCTBuilder::above(gemmInputTransform, gemmInputTransformAttr);
-
-      padGemmInputTransform.passThrough("gemmG");
-      if (inputCheckPadGemmK) {
-        padGemmInputTransform.pad("gemmKPad", "gemmK", 0, gemmKExtra);
-
-        // Forward convolution has a K made from c and parts of w and h
-        // while backward weights have n and parts of w and h
-        inputOobCheckDims.insert(padInputTransform.startIndex("hi"));
-        inputOobCheckDims.insert(padInputTransform.startIndex("wi"));
-
-        inputOobCheckDims.insert(padInputTransform.startIndex(
-            convOpType == miopen::ConvOpType::Fwd ? "ci" : "ni"));
-      } else {
-        padGemmInputTransform.passThrough("gemmK");
-      }
-
-      if (inputCheckPadGemmN) {
-        padGemmInputTransform.pad("gemmNPad", "gemmN", 0, gemmNExtra);
-
-        // Forward convolution has a N made from n and parts of w and h
-        // while backward weights have c and parts of w and h
-        inputOobCheckDims.insert(padInputTransform.startIndex("hi"));
-        inputOobCheckDims.insert(padInputTransform.startIndex("wi"));
-
-        inputOobCheckDims.insert(padInputTransform.startIndex(
-            convOpType == miopen::ConvOpType::Fwd ? "ni" : "ci"));
-      } else {
-        padGemmInputTransform.passThrough("gemmN");
-      }
-
-      TransformMapAttr padGemmInputTransformAttr = padGemmInputTransform.get();
-      gemmInputPad = b.create<miopen::TransformOp>(loc, gemmInput,
-                                                   padGemmInputTransformAttr);
-      // input padding end
-    }
-
-    // Transform output tensor.
-    OobCheckSet outputOobCheckDims;
-    // - PassThrough G to dimmension 0, name it gemmG, then
-    // Output tensor transformation for Conv2DOp:
-    // - PassThrough K dimension to dimension 1, named gemmM
-    // - Merge non-K dimensions to dimension2, named gemmN
-
-    // Output tensor transformation for backward weight:
-    // - Merge non-K dimensions to dimension 1, named gemmK
-    // - PassThrough K dimension to dimension 2, name it gemmM
-    SmallVector<StringRef, 5> outputNonKDims;
-    for (StringRef name : outputNames)
-      if (name != "go" && name != "ko")
-        outputNonKDims.push_back(name);
-
-    BottomUpCTBuilder outputTransform(b, outputNames, outputShape, loc);
-    outputTransform.passThrough({"gemmG"}, {0}, {"go"});
-    switch (convOpType) {
-    case miopen::ConvOpType::Fwd:
-      outputTransform.passThrough({"gemmM"}, {1}, {"ko"});
-      outputTransform.merge("gemmN", 2, outputNonKDims);
-      break;
-    case miopen::ConvOpType::BwdWeight:
-      outputTransform.merge("gemmK", 1, outputNonKDims);
-      outputTransform.passThrough({"gemmM"}, {2}, {"ko"});
-      break;
-    case miopen::ConvOpType::BwdData:
-      llvm_unreachable("Backward data has been sent elsewhere");
-      break;
-    }
-
-    TransformMapAttr outputTransformAttr = outputTransform.get();
-    Value gemmOutput =
-        b.create<miopen::TransformOp>(loc, op.output(), outputTransformAttr);
-    Value gemmOutputPad = gemmOutput;
-
-    // output padding start
-    // output matrix dim: K & NHW
-    // when backward weight , GEMMK = NHW
-    // N:batch size, H:output height ,W:output width
-    // If size of N*h*w is under 32 or 64 ,we pad it to 32 or 64
-    // then mlir can do gemm
-    // we just add more one transform to do it
-
-    bool outputCheckPadGemmK = false;
-    bool outputCheckPadGemmM = false;
-    bool outputCheckPadGemmN = false;
-    outputCheckPadGemmK =
-        (convOpType == miopen::ConvOpType::BwdWeight && gemmKExtra > 0);
-    outputCheckPadGemmM =
-        (convOpType == miopen::ConvOpType::BwdWeight && gemmMExtra > 0) ||
-        (convOpType == miopen::ConvOpType::Fwd && gemmMExtra > 0);
-    outputCheckPadGemmN =
-        (convOpType == miopen::ConvOpType::Fwd && gemmNExtra > 0);
-    if (outputCheckPadGemmK || outputCheckPadGemmM || outputCheckPadGemmN) {
-      auto padGemmOutputTransform =
-          BottomUpCTBuilder::above(outputTransform, outputTransformAttr);
-      padGemmOutputTransform.passThrough("gemmG");
-
-      if (outputCheckPadGemmK) {
-        padGemmOutputTransform.pad("gemmKPad", "gemmK", 0, gemmKExtra);
-        // Unlike in the filter case, add all the dimensions to the oob check
-        // since this is loading data during backward weights, not storing it
-        outputOobCheckDims.insert(outputTransform.startIndex("no"));
-        outputOobCheckDims.insert(outputTransform.startIndex("ho"));
-        outputOobCheckDims.insert(outputTransform.startIndex("wo"));
-      } else if (convOpType != miopen::ConvOpType::Fwd) {
-        padGemmOutputTransform.passThrough("gemmK");
-      }
-
-      if (outputCheckPadGemmM) {
-        padGemmOutputTransform.pad("gemmMPad", "gemmM", 0, gemmMExtra);
-        // For the cases considered in this function, the m dimension of the
-        // gemm is always the k dimension of the output tensor
-        outputOobCheckDims.insert(outputTransform.startIndex("ko"));
-      } else {
-        padGemmOutputTransform.passThrough("gemmM");
-      }
-
-      if (outputCheckPadGemmN) {
-        padGemmOutputTransform.pad("gemmNPad", "gemmN", 0, gemmNExtra);
-        // As in the filter case with backward weight, set only the outermost
-        // dimension of the output tensor for OOB checks or else extra 0s will
-        // appear in the output
-        outputOobCheckDims.insert(
-            outputTransform.startIndex(outputNonKDims[0]));
-      } else if (convOpType != miopen::ConvOpType::BwdWeight) {
-        padGemmOutputTransform.passThrough("gemmN");
-      }
-
-      TransformMapAttr padGemmOutputTransformAttr =
-          padGemmOutputTransform.get();
-      gemmOutputPad = b.create<miopen::TransformOp>(loc, gemmOutput,
-                                                    padGemmOutputTransformAttr);
-      // output padding end
-    }
-
-    // Set attributes for gridwise_gemm op.
-    llvm::SmallVector<NamedAttribute, 8> gridwiseGemmAttrs{
-        b.getNamedAttr("arch", archAttr),
-        b.getNamedAttr("num_cu", numCuAttr),
-    };
-
-    // xdlopsV2.
-    if (isXdlops)
-      gridwiseGemmAttrs.push_back(
-          b.getNamedAttr("xdlopsV2", b.getBoolAttr(true)));
-
-    if (convOpType == miopen::ConvOpType::BwdData) {
-      gridwiseGemmAttrs.push_back(b.getNamedAttr(
-          "kernel_algorithm", b.getStringAttr("backward_data_v1r1")));
-    } else if (convOpType == miopen::ConvOpType::Fwd) {
-      gridwiseGemmAttrs.push_back(
-          b.getNamedAttr("kernel_algorithm", b.getStringAttr("v4r4")));
-    } else if (convOpType == miopen::ConvOpType::BwdWeight) {
-      gridwiseGemmAttrs.push_back(b.getNamedAttr(
-          "kernel_algorithm", b.getStringAttr("backward_weight_v4r4")));
-    }
-
-    // Gather up OOB check attribute arrays
-    ArrayAttr filterOobAttr =
-        getBoundsCheckAttr(b, filterOobCheckDims, filterShape.size());
-    ArrayAttr inputOobAttr =
-        getBoundsCheckAttr(b, inputOobCheckDims, inputShape.size());
-    ArrayAttr outputOobAttr =
-        getBoundsCheckAttr(b, outputOobCheckDims, outputShape.size());
-
-    SmallVector<Value, 3> arguments = {gemmFilterPad, gemmInputPad,
-                                       gemmOutputPad};
-    SmallVector<ArrayAttr, 3> oobs = {filterOobAttr, inputOobAttr,
-                                      outputOobAttr};
-
-    Value gemmA, gemmB, gemmC;
-    ArrayAttr oobA, oobB, oobC;
-    gemmA = arguments[fields.gridwiseGemmArgumentPosition[0]];
-    oobA = oobs[fields.gridwiseGemmArgumentPosition[0]];
-    gemmB = arguments[fields.gridwiseGemmArgumentPosition[1]];
-    oobB = oobs[fields.gridwiseGemmArgumentPosition[1]];
-    gemmC = arguments[fields.gridwiseGemmArgumentPosition[2]];
-    oobC = oobs[fields.gridwiseGemmArgumentPosition[2]];
-
-    // Create padding info attr
-    PaddingInfoAttr paddingInfo =
-        PaddingInfoAttr::get(b.getContext(), gemmMExtra, gemmKExtra, gemmNExtra,
-                             BwdPaddingKernelInfo::NA);
-
-    auto dataOp = InMemoryDataOperation::Set;
-    // Emit miopen.gridwise_gemm op.
-    // Emit miopen.gridwise_gemm_v2 if xdlopsV2 attribute is true.
-
-    if (xdlopsV2Attr && xdlopsV2Attr.getValue() == true) {
-      auto gop = b.create<miopen::GridwiseGemmV2Op>(
-          loc, gemmA, gemmB, gemmC, oobA, oobB, oobC, paddingInfo, dataOp,
-          gridwiseGemmAttrs);
-      affixGridwiseGemmAttributes(op, gop, b);
-    } else {
-      auto gop = b.create<miopen::GridwiseGemmOp>(loc, gemmA, gemmB, gemmC,
-                                                  oobA, oobB, oobC, paddingInfo,
-                                                  gridwiseGemmAttrs);
-      affixGridwiseGemmAttributes(op, gop, b);
-    }
-
-    // Finally, erase the original Conv2D op.
-    op.erase();
-
-    return success();
-  }
-};
-
-// Forward-declare field values to suppress warnings
-template <> const ArgumentFields Conv2DRewritePattern<miopen::Conv2DOp>::fields;
-template <>
-const miopen::ConvOpType Conv2DRewritePattern<miopen::Conv2DOp>::convOpType;
-template <>
-const ArgumentFields Conv2DRewritePattern<miopen::Conv2DBwdDataOp>::fields;
-template <>
-const miopen::ConvOpType
-    Conv2DRewritePattern<miopen::Conv2DBwdDataOp>::convOpType;
-template <>
-const ArgumentFields Conv2DRewritePattern<miopen::Conv2DBwdWeightOp>::fields;
-template <>
-const miopen::ConvOpType
-    Conv2DRewritePattern<miopen::Conv2DBwdWeightOp>::convOpType;
 //===----------------------------------------------------------------------===//
 // Assigning attributes.
 //===----------------------------------------------------------------------===//
